+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/amazon-bedrock
 
+## 5.0.27
+
+### Patch Changes
+
+- Updated dependencies [97de198]
+  - @ai-sdk/anthropic@4.0.18
+
 ## 5.0.26
 
 ### Patch Changes